Move HeaderErrorWidget
