+++ lib/ui/header_error_widget.dart
